DEV-01 Cleaning, Connecting tmp mysql

+++ src/main/resources/db/migration/V1_0__INIT_TABLE_BANK_PROVIDER.sql
@@ -4,7 +4,7 @@ CREATE TABLE bank_provider (
   name char(32)  NOT NULL,
   overdraft integer  NOT NULL,
   incentive char(100)  NOT NULL,
-  student_score float(100)  NOT NULL,
+  student_score float(5, 2)  NOT NULL,
   location_id integer  NULL,
   CONSTRAINT bank_provider_pk PRIMARY KEY (id)
 ) ;
